+++ src/API/LeadershipProfileAPI/Features/Search/List.cs
@@ -39,7 +39,7 @@ namespace LeadershipProfileAPI.Features.Search
             public string MiddleName { get; set; }
             public string LastSurName { get; set; }
             public string FullName { get; set; }
-            public int YearsOfService { get; set; }
+            public decimal YearsOfService { get; set; }
             public string Certification { get; set; }
             public string Assignment { get; set; }
             public string Degree { get; set; }
